♻️ Change flag opening/closing logic
There is the issue with using CSS for opening/closing logic for flags. It's broken on touch devices (which don't trigger the `:hover` styling) (Issue #77). To fix it, we moved all the logic to JS:
- Changed `z-index` to «move» a cursor behind the text. It resolves the issue, when a cursor prevent clicking on the text (even if a user clicks exactly on cursor)
- Added listeners for `mousemove` and `touchstart` with a callback, which shows/closes flags depends on a point position

To achieve it there are added a couple of callbacks to update cursors position and store it in the quill-cursors module. It should be enough to cover all cases for touch and non-touch devices. Also, there is a small throttling for mousemove events to improve performance and decrease amount of callbacks executions